+++ microraidenj/src/main/java/com/asf/microraidenj/eth/interfaces/GetNonce.java
@@ -1,8 +1,9 @@
 package com.asf.microraidenj.eth.interfaces;
 
 import com.asf.microraidenj.type.Address;
+import java.math.BigInteger;
 
 public interface GetNonce {
 
-  long get(Address address);
+  BigInteger get(Address address);
 }
