+++ microraidenj/src/main/java/com/asf/microraidenj/eth/interfaces/TransactionSender.java
@@ -3,9 +3,10 @@ package com.asf.microraidenj.eth.interfaces;
 import com.asf.microraidenj.exception.TransactionFailedException;
 import com.asf.microraidenj.type.Address;
 import ethereumj.crypto.ECKey;
+import java.math.BigInteger;
 
 public interface TransactionSender {
 
-  String send(ECKey senderECKey, Address receiveAddress, long value, byte[] data)
+  String send(ECKey senderECKey, Address receiveAddress, BigInteger value, byte[] data)
       throws TransactionFailedException;
 }
